# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AURA-OMEGA - Multi-agent orchestrator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vault:
    get:
      operationId: listVaultSecrets
      tags: [vault]
      summary: List stored secrets (metadata only, values never returned)
      responses:
        "200":
          description: List of secret metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultSecret"
        "401":
          description: Operator sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: setVaultSecret
      tags: [vault]
      summary: Create or update a secret (encrypted at rest)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetVaultSecretRequest"
      responses:
        "200":
          description: Stored secret metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultSecret"
        "400":
          description: Invalid secret data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Operator sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /vault/{name}:
    delete:
      operationId: deleteVaultSecret
      tags: [vault]
      summary: Delete a secret by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteVaultSecretResult"
        "401":
          description: Operator sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    VaultSecret:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SetVaultSecretRequest:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
        description:
          type: string
    DeleteVaultSecretResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: string
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string